Refactored the partners section out
