--- src/components/ExchangePage/logos/bigmama.js
@@ -1,16 +1,0 @@
-// @flow
-
-import React from 'react'
-
-const inner = (
-  <g fill="#135577" fillRule="nonzero">
-    <path d="M22.651 14.519c.018.27.203.471.314.702.149.312.334.606.503.908.18.318.36.636.535.954.23.383.434.781.61 1.192.072.18.151.353.215.533.127.36.044.604-.207.908-.009.01-.026.015-.033.026-.203.305-.525.36-.854.42-.29.055-.58.126-.797.364-.216.238-.246.438-.061.713.185.275.36.539.541.815a.76.76 0 0 1-.217 1.085c-.15.091-.295.192-.433.3-.114.09-.107.119-.02.245.196.283.37.574.375.935a.565.565 0 0 1-.288.528c-.296.198-.636.277-.951.42-.415.188-.586.5-.307.923.272.386.376.865.289 1.328a1.97 1.97 0 0 1-.242.66 2.95 2.95 0 0 1-.631.77c-.15.131-.33.225-.523.273-.195.042-.38.14-.575.174-.502.09-.992.233-1.508.27-.373.025-.74.105-1.111.15-.221.027-.447.022-.669.049-.824.098-1.653.067-2.48.095-.263.01-.516.1-.771.18-.24.072-.427.224-.631.342a.604.604 0 0 0-.203.215c-.1.17-.18.352-.27.528l-.357.67c-.117.222-.235.473-.355.71-.26.515-.525 1.028-.785 1.545-.16.318-.317.639-.477.958-.23.46-.462.918-.696 1.376a.427.427 0 0 1-.142.17 3324.64 3324.64 0 0 1-.04-3.454c-.015-1.155-.032-2.309-.05-3.502.12-.008.22-.019.318-.02h.758c.286 0 .295-.009.295-.286v-2.188c.14-.11.293-.083.43-.083 1.024-.005 2.049 0 3.074 0 .185 0 .371.018.508.026.129.188.27.368.421.538.354.343.932.48 1.407.219a1.51 1.51 0 0 0 .616-.604c.207-.372.103-.968-.174-1.297a1.474 1.474 0 0 0-.614-.434 1.066 1.066 0 0 0-.707-.006 1.434 1.434 0 0 0-.69.474c-.074.096-.13.203-.169.317a.24.24 0 0 1-.218.15c-.155.015-.313.01-.47.011H9.587c-.244 0-.265-.019-.27-.251 0-.388-.01-.776-.01-1.168 0-.348 0-.35.345-.35.505 0 1.01.005 1.515 0 .09 0 .18-.015.292-.025v-.254-2.306c0-.18-.005-.192-.195-.193-.57 0-1.14 0-1.713-.004H9.28V20.82c.114-.008.226-.02.34-.02h6.342c.548 0 1.096-.01 1.644-.023a.328.328 0 0 1 .343.193c.098.19.238.355.41.483.3.196.647.303 1.008.19.23-.061.441-.173.622-.326.39-.359.54-1.178.134-1.616-.123-.135-.216-.295-.387-.384a1.327 1.327 0 0 0-1.134-.081c-.285.107-.46.326-.639.538-.064.078-.078.196-.116.3-.156.043-5.459.06-5.741.011-.007-.099-.018-.204-.018-.309V17.19v-.13c-.01-.194-.036-.222-.234-.222H9.262c-.005-.098-.013-.175-.013-.252 0-.366 0-.733-.005-1.099-.005-.343-.023-.686-.033-1.03v-.23c.095-.004.173-.01.25-.01h6.56c.163 0 .286.018.352.199.072.197.245.328.392.471.077.06.164.106.257.134.478.193.902.112 1.304-.216.312-.248.493-.625.49-1.023a.97.97 0 0 0-.18-.54c-.24-.386-.35-.473-.804-.64a.884.884 0 0 0-.611 0c-.054.02-.11.03-.164.048a.669.669 0 0 0-.313.18.685.685 0 0 1-.138.09c-.023.016-.055.026-.067.049-.105.18-.207.365-.315.558h-5.04c-.008-.1-.018-.18-.018-.251v-1.25c0-.242-.04-.28-.289-.28H9.233c-.064-.279-.077-2.192-.013-2.39.075 0 .16-.01.243-.01h3.355c.203-.009.24-.043.24-.242V7.566a.933.933 0 0 1 .155-.029h1.3a.245.245 0 0 1 .224.127c.11.156.18.339.338.466.373.3.79.436 1.247.276.253-.09.504-.231.642-.493a.706.706 0 0 1 .06-.09c.252-.348.254-.718.108-1.104a1.168 1.168 0 0 0-.439-.561 1.345 1.345 0 0 0-.666-.27 1.238 1.238 0 0 0-.679.151c-.347.188-.55.475-.69.841H12.24c-.006-1.06-.01-2.11 0-3.182.147-.008.27-.019.389-.019h2.51c.051.003.102.003.152 0 .135-.012.16-.038.166-.173V3.29.022c.108-.007.198-.018.289-.018.664 0 1.328.011 1.99 0 .194.001.383.06.542.171.243.158.47.338.679.539.261.238.506.494.732.765.209.257.424.508.612.781.187.274.36.54.413.874.042.23.097.456.167.678a.48.48 0 0 0 .12.174c.083.09.172.172.261.254.185.185.348.39.488.611.176.245.332.503.465.773.148.303.26.628.38.942.098.257.18.517.27.777a.405.405 0 0 1 .023.103c.02.314.052.628.055.944.011 1.105.014 2.212.015 3.317 0 .18 0 .36-.133.525-.061.072-.063.2-.077.306-.03.219-.012.454-.086.656-.097.263-.05.521-.06.78M5.064 29.16v-3.335c.155-.043 3.043-.06 3.387-.02.047.143.061 3.048.014 3.349-.063.006-.131.02-.2.021H5.28c-.064.004-.127-.006-.216-.014zM4.34 11.361V8.36c.029-.006.055-.021.07-.014.18.09.378.06.57.062h2.164c.078 0 .155.007.246.011v2.933c-.17.06-2.825.069-3.05.01z" />
-    <path d="M14.3 2.852h-2.673c-.056-.171-.073-2.365-.024-2.67.147-.047 2.33-.065 2.697-.017.003.888 0 1.78 0 2.687zM11.04 4.513v2.502c-.163.048-2.264.062-2.554.015V4.515l2.555-.002zM7.01 16.742H4.705c-.053-.157-.066-2.1-.018-2.353.164-.043 2.105-.05 2.321-.01.052.19.056 2.161.002 2.364zM7.306 18.878v2.03c-.146.054-1.81.07-2.086.023-.051-.157-.065-1.792-.017-2.05.168-.043 1.865-.05 2.103-.003zM7.885 31.1v1.922c-.09.006-.166.015-.243.016h-1.21c-.164 0-.328.019-.486.028-.064-.158-.09-1.668-.034-1.956.152-.047 1.691-.059 1.973-.01zM.826 24.565v-1.862h1.88c.052.18.062 1.657.01 1.862H.827zM8.76 17.989h1.726c.053.155.063 1.552.01 1.79H8.77c-.02-.593-.01-1.184-.008-1.79zM.034 13.56H1.56v1.521a.724.724 0 0 1-.145.029H.182c-.151 0-.174-.017-.176-.153C0 14.532 0 14.11 0 13.687a.58.58 0 0 1 .034-.126zM2.15 17.635v1.494H.606a1.632 1.632 0 0 1-.018-.164v-1.167c0-.172.016-.197.18-.199.418-.005.836 0 1.255 0 .043.008.085.02.125.036zM7.306 22.272v1.437c-.162.064-1.234.07-1.469.009v-1.446h1.469zM40.543 27.885c-1.407 0-2.67-.216-3.788-.647a7.479 7.479 0 0 1-2.842-1.9c-.777-.834-1.372-1.852-1.786-3.053-.414-1.2-.621-2.583-.622-4.149 0-1.562.239-2.945.717-4.149a8.964 8.964 0 0 1 1.961-3.053 8.238 8.238 0 0 1 2.923-1.9 9.916 9.916 0 0 1 3.598-.646c.708-.004 1.414.056 2.111.18.557.094 1.108.22 1.65.377.402.117.791.27 1.164.458.307.161.522.278.65.35l-1.002 2.775c-.433-.27-1.05-.525-1.853-.768a8.997 8.997 0 0 0-2.611-.36 6.121 6.121 0 0 0-2.273.418 4.919 4.919 0 0 0-1.827 1.253 5.816 5.816 0 0 0-1.204 2.102c-.288.844-.433 1.823-.433 2.936-.009.918.105 1.833.339 2.721.2.777.554 1.505 1.041 2.142a4.73 4.73 0 0 0 1.786 1.401c.722.333 1.588.499 2.598.499 1.207 0 2.19-.12 2.949-.36a11.786 11.786 0 0 0 1.759-.695l.92 2.775a4.726 4.726 0 0 1-.88.43 12.77 12.77 0 0 1-3.017.741c-.672.084-1.35.124-2.028.122zM60.404 20.422a9.06 9.06 0 0 1-.487 3.044 6.73 6.73 0 0 1-1.381 2.344 6.204 6.204 0 0 1-2.151 1.509 6.971 6.971 0 0 1-2.774.539 6.87 6.87 0 0 1-2.76-.54 6.236 6.236 0 0 1-2.137-1.508 6.9 6.9 0 0 1-1.394-2.344 8.818 8.818 0 0 1-.5-3.044 8.704 8.704 0 0 1 .5-3.031 6.796 6.796 0 0 1 1.407-2.335 6.202 6.202 0 0 1 2.151-1.495 6.951 6.951 0 0 1 2.733-.526 7.062 7.062 0 0 1 2.747.526 6.055 6.055 0 0 1 2.15 1.495 6.94 6.94 0 0 1 1.394 2.335c.345.973.515 1.999.502 3.03zm-3.356 0c0-1.401-.3-2.51-.902-3.327-.6-.818-1.446-1.226-2.534-1.226-1.082 0-1.924.408-2.525 1.226-.602.817-.905 1.926-.911 3.327 0 1.419.302 2.541.906 3.368.604.826 1.446 1.239 2.526 1.239 1.082 0 1.925-.413 2.53-1.24.604-.826.907-1.948.91-3.367zM66.085 9.268c0 .61-.198 1.096-.595 1.455a2.027 2.027 0 0 1-1.407.539 2.062 2.062 0 0 1-1.434-.54c-.397-.358-.596-.843-.596-1.454 0-.629.199-1.123.596-1.482.39-.357.904-.55 1.434-.539a2.027 2.027 0 0 1 1.407.54c.397.358.595.852.595 1.481zm-.352 18.213H62.46V13.39h3.274v14.09zM68.547 13.848c.81-.217 1.633-.388 2.463-.512a21.32 21.32 0 0 1 3.355-.242c1.154 0 2.12.157 2.895.471.776.314 1.393.755 1.853 1.32.477.6.81 1.3.975 2.048.196.862.291 1.743.284 2.627v7.92h-3.274v-7.408c.008-.646-.041-1.29-.15-1.927-.071-.46-.237-.9-.486-1.293a1.943 1.943 0 0 0-.92-.727 3.894 3.894 0 0 0-1.42-.23 10.96 10.96 0 0 0-1.3.081c-.45.054-.784.1-1 .135v11.37h-3.275V13.848zM83.267 14.037a38.228 38.228 0 0 1 2.043-.485c.839-.18 1.925-.27 3.26-.27 1.768 0 3.049.512 3.842 1.536a7.407 7.407 0 0 1 1.52-.89 7.695 7.695 0 0 1 3.166-.646c1.064 0 1.926.157 2.584.472a3.61 3.61 0 0 1 1.529 1.32c.38.62.63 1.313.73 2.034.131.856.194 1.72.19 2.586v7.787h-1.759v-7.22a20.238 20.238 0 0 0-.122-2.385c-.08-.674-.243-1.23-.487-1.67-.235-.432-.599-.78-1.041-.997-.451-.225-1.038-.337-1.76-.337a6.773 6.773 0 0 0-1.474.148c-.394.083-.78.205-1.15.364a4.86 4.86 0 0 0-.798.43 12.14 12.14 0 0 0-.451.324c.173.512.295 1.04.365 1.576.083.612.124 1.228.122 1.846v7.92h-1.759v-7.22a20.238 20.238 0 0 0-.122-2.384c-.08-.674-.248-1.23-.5-1.67a2.44 2.44 0 0 0-1.069-.997c-.46-.225-1.069-.337-1.826-.337a14.5 14.5 0 0 0-2.151.135c-.587.09-.961.17-1.123.242v12.232h-1.76V14.037zM109.512 13.228c.92 0 1.702.13 2.345.391.598.23 1.131.6 1.556 1.078.405.473.701 1.03.866 1.63.186.666.277 1.356.27 2.047v8.783c-.258.073-.52.133-.784.18-.343.06-.74.127-1.19.202a17.21 17.21 0 0 1-1.502.175c-.55.045-1.114.067-1.691.067a8.115 8.115 0 0 1-2.043-.242 4.524 4.524 0 0 1-1.624-.768 3.626 3.626 0 0 1-1.082-1.347c-.262-.548-.393-1.217-.393-2.007 0-.755.145-1.401.433-1.94a3.675 3.675 0 0 1 1.218-1.334 5.474 5.474 0 0 1 1.867-.767 10.765 10.765 0 0 1 2.354-.243c.253 0 .519.014.798.04a11.078 11.078 0 0 1 1.488.243c.135.027.266.068.393.121v-.862c0-.473-.035-.946-.108-1.414a3.09 3.09 0 0 0-.474-1.24 2.625 2.625 0 0 0-1.028-.889c-.442-.224-1.024-.336-1.745-.336-1.029 0-1.795.071-2.3.215-.505.144-.875.26-1.11.35l-.243-1.535c.306-.144.78-.279 1.42-.405a12.201 12.201 0 0 1 2.31-.193zm.163 12.986c.685 0 1.286-.022 1.804-.067.44-.035.878-.093 1.312-.175v-4.904a5.411 5.411 0 0 0-.947-.282c-.433-.1-1.028-.149-1.786-.149-.453 0-.904.03-1.353.09a3.917 3.917 0 0 0-1.299.404c-.389.2-.723.491-.974.849-.252.36-.379.835-.379 1.428a3.05 3.05 0 0 0 .257 1.32c.162.348.414.645.731.862.345.228.73.388 1.136.472.493.105.995.156 1.499.152h-.001zM117.684 14.037a38.228 38.228 0 0 1 2.043-.485c.839-.18 1.925-.27 3.26-.27 1.768 0 3.049.512 3.842 1.536a7.46 7.46 0 0 1 1.52-.89 7.695 7.695 0 0 1 3.166-.646c1.064 0 1.926.157 2.584.472.627.285 1.16.743 1.533 1.32.381.62.63 1.313.731 2.034.131.856.194 1.72.19 2.586v7.787h-1.76v-7.22a20.238 20.238 0 0 0-.121-2.385c-.081-.674-.244-1.23-.487-1.67a2.373 2.373 0 0 0-1.042-.997c-.45-.225-1.037-.337-1.759-.337a6.773 6.773 0 0 0-1.474.148c-.395.084-.78.205-1.15.364a4.899 4.899 0 0 0-.799.43 12.14 12.14 0 0 0-.45.324c.17.513.292 1.04.36 1.576.083.612.124 1.228.122 1.846v7.92h-1.759v-7.22a20.238 20.238 0 0 0-.122-2.384c-.08-.674-.248-1.23-.5-1.67a2.436 2.436 0 0 0-1.069-.997c-.46-.225-1.069-.337-1.826-.337-.72-.009-1.439.036-2.151.135-.587.09-.961.17-1.123.242v12.232h-1.76V14.037zM143.93 13.228c.92 0 1.702.13 2.345.391.598.23 1.131.6 1.556 1.078.405.473.7 1.03.866 1.63.185.666.277 1.356.27 2.047v8.783c-.258.073-.52.133-.784.18-.343.06-.74.127-1.19.202a16.19 16.19 0 0 1-1.51.17c-.55.045-1.114.068-1.692.068a8.115 8.115 0 0 1-2.043-.243 4.524 4.524 0 0 1-1.623-.767 3.617 3.617 0 0 1-1.082-1.347c-.262-.548-.393-1.217-.393-2.008 0-.754.145-1.4.433-1.94a3.66 3.66 0 0 1 1.218-1.333 5.474 5.474 0 0 1 1.867-.768 10.759 10.759 0 0 1 2.354-.242c.252 0 .518.013.798.04a11.078 11.078 0 0 1 1.488.243c.135.027.266.067.393.12v-.861c0-.474-.036-.947-.109-1.415a3.09 3.09 0 0 0-.473-1.24 2.625 2.625 0 0 0-1.028-.888c-.442-.225-1.024-.337-1.746-.337-1.028 0-1.794.072-2.3.216-.505.143-.874.26-1.109.35l-.243-1.536c.306-.144.78-.278 1.42-.404.765-.137 1.54-.2 2.317-.189zm.163 12.986c.685 0 1.286-.022 1.803-.067.44-.035.879-.093 1.313-.175v-4.904a5.411 5.411 0 0 0-.947-.282c-.433-.1-1.028-.149-1.786-.149-.453 0-.905.03-1.353.09a3.91 3.91 0 0 0-1.299.404c-.389.2-.723.491-.974.849-.253.36-.38.835-.379 1.428-.013.453.074.904.257 1.32.162.348.414.645.73.862.346.228.732.388 1.137.472.492.105.994.156 1.498.152z" />
-  </g>
-)
-
-export default ({ width }: { width: number }) => (
-  <svg width={width} viewBox="0 0 149 37">
-    {inner}
-  </svg>
-)
