--- src/components/ExchangePage/logos/coinberry.js
@@ -1,116 +1,0 @@
-// @flow
-
-import React from 'react'
-
-const styles = {
-  blue: {
-    fill: '#334F93',
-  },
-  pink: {
-    fill: '#EC2D6E',
-  },
-}
-
-const inner = (
-  <>
-    <path
-      style={styles.blue}
-      d="M146.3,101.8c5.2,0,9.4-4,9.4-9.4c0-5.2-4.1-9.5-9.4-9.5c-5.5,0-9.5,4.4-9.5,9.5
-	C136.8,97.7,140.7,101.8,146.3,101.8"
-    />
-    <path
-      style={styles.blue}
-      d="M102.3,164.7c8.4,0,9.4-10.3,9.4-25.7c0-18.7-2-25.8-9.5-25.8c-8.9,0-9.6,12.1-9.6,25.8
-	C92.6,152.1,92.6,164.7,102.3,164.7 M102.2,109.4c15.9,0,27.3,12.7,27.3,29.6c0,16.1-10.8,29.5-27.3,29.5
-	c-16.9,0-27.4-13.7-27.4-29.5C74.8,123.4,85.4,109.4,102.2,109.4"
-    />
-    <path
-      style={styles.blue}
-      d="M246.2,155c0,3.4,1.6,9.6,8.3,9.6c7.3,0,11.5-6.1,11.5-25.2c0-7.1-0.8-23.3-11.3-23.3c-5.6,0-7.4,4.5-8.5,7.1
-	V155z M222.6,84.9h23.6v31.4c2-2.2,6.5-6.8,14.7-6.8c14,0,22.8,12.3,22.8,28.1c0,15.4-8.3,30.9-28.3,30.9c-2.8,0-6.8-0.9-11-3.5
-	c-2.8-1.8-6.3-1.8-9,0.1l-4.7,3.3h-1.4V94.8c0-5.7-2.3-6.1-6.7-7V84.9z"
-    />
-    <path
-      style={styles.blue}
-      d="M77.8,137.3H76c-0.2,6.7-2.4,12.8-5.9,16.9c-0.6,0.8-2.6,2.6-3.5,3.4c-2,1.7-4.8,2.8-8.8,2.8
-	c-9.5,0-16.9-9.8-16.9-26.8c0-5.5,0.9-20.3,10.3-20.3c1.8,0,4.9,1.1,4.9,4.7c0,6.6,0,12.5,6.8,12.5c1,0,8.3,0,8.3-7.7
-	c0-8.7-10.2-13.2-18.8-13.2c-11.4,0-28.6,7.4-28.6,31.3c0,15,10.4,27.8,25.9,27.8c7.6,0,14.3-3,19.5-8.3
-	C70.2,159.2,77.8,152.3,77.8,137.3"
-    />
-    <path
-      style={styles.blue}
-      d="M370.1,166.7h-32.3v-2.9c3.9-0.8,6.6-1.2,6.6-7.3V121c0-6-2.9-6.7-6.6-7.1v-3h22.9l0.1,4.4c0,1.6,2,2.4,3.2,1.1
-	c3-3.4,7.3-7,12.8-7c8.3,0,10.1,6.6,10.1,9.5c0,3.1-1.7,8.1-7.7,8.1c-7,0-6.9-7.1-10.8-7.1c-1.8,0-7.1,2.7-7.1,11.5v22.9
-	c0,7.4,1.5,9.3,8.8,9.6v2.9H370.1z"
-    />
-    <path
-      style={styles.blue}
-      d="M218.1,157v-30.4c0-13.3-9.6-17.1-16.2-17.1c-6.6,0-11.9,4.2-14.8,7.2c-1,1-2.6,0.3-2.6-1.1v-4.6h-23.1v2.9
-	c4.8,0.8,6.5,1.7,6.5,7.1v35.6v0.3v1.8c0,2.8-2.3,5.1-5.1,5.1H160c-2.8,0-5.1-2.3-5.1-5.1v-24.9v-22.9h-23.2v2.9
-	c4.6,0.9,6.5,1.7,6.5,7.1v35.6c0,5.8-2,6.1-6.5,7.3v2.9h27.4h3.6h28v-2.9c-5.6-0.6-5.9-4.1-5.9-6.9v-32.4c0-1.1,5.2-7,10.1-7
-	c6.6,0,6.7,6.7,6.7,10.1v29.4c0,4.5-1.7,6.6-5.7,6.9v2.9h28.5V164C218.1,163.6,218.1,159.9,218.1,157"
-    />
-    <path
-      style={styles.blue}
-      d="M475.2,110.9v2.9c3,0.1,6.7,0.1,6.7,4c0,1.8-0.6,3.4-1.3,5.5l-8.3,23.2l-8.7-22.4c-1.1-2.7-2.5-5.8-2.5-7
-	c0-2.9,3.2-3,6.5-3.2v-3H446c-3,0-5,3.1-3.8,5.8l7.2,15.9l14.5,35.7l-2.2,6.5c-0.9,2.7-3.9,9.9-8,9.9c-1.5,0-2-1.5-2-2
-	c0-0.5,0.3-1.1,0.3-2.2c0-2.2-1.3-6.1-6.8-6.1c-7.3,0-8,6.1-8,7.9c0,3.9,3.2,9.5,11.2,9.5c11.4,0,14-6.9,20-23l18.1-48.1
-	c2.3-6,4.4-6.3,7-6.7v-3H475.2z"
-    />
-    <path
-      style={styles.blue}
-      d="M420.9,166.7h-32.3v-2.9c3.9-0.8,6.6-1.2,6.6-7.3V121c0-6-2.9-6.7-6.6-7.1v-3h22.9l0.1,4.4c0,1.6,2,2.4,3.2,1.1
-	c3-3.4,7.3-7,12.8-7c8.3,0,10.1,6.6,10.1,9.5c0,3.1-1.7,8.1-7.7,8.1c-7,0-6.9-7.1-10.8-7.1c-1.8,0-7.1,2.7-7.1,11.5v22.9
-	c0,7.4,1.5,9.3,8.8,9.6v2.9H420.9z"
-    />
-    <path
-      style={styles.blue}
-      d="M487.5,159.5h0.9c1,0,1.4-0.8,1.4-1.8c0-1.2-0.6-1.5-1.4-1.5h-0.9V159.5z M485.2,155.8h3.5
-	c1.2,0,2.4,0.5,2.4,1.8c0,1.1-0.6,1.8-1.6,2.2l1.2,1.6c0.5,0.8,1.2,1.4,1.5,1.7v0.2h-1.3c-0.6,0-1.2-1.4-2.5-3.2h-1v2
-	c0,0.9,0,0.9,1,1v0.3h-3.2v-0.3c1-0.1,1-0.1,1-1v-4.8c0-0.9,0-0.9-1-1V155.8z M488.6,153.4c-3.3,0-5.8,2.9-5.8,6.1s2.6,6.1,5.8,6.1
-	c3.2,0,5.8-2.9,5.8-6.1C494.3,156.3,491.8,153.4,488.6,153.4 M488.6,166.5c-3.9,0-7-3.1-7-7c0-3.9,3-7,7-7c3.9,0,6.9,3.1,6.9,7
-	C495.5,163.4,492.4,166.5,488.6,166.5"
-    />
-    <path
-      style={styles.blue}
-      d="M321.8,132.6c-0.1-3.8-0.1-10.1-1.1-13.7c-0.8-2.7-2.3-5.7-6.6-5.7c-4.5,0-8.6,2.7-8.6,19.4H321.8z
-	 M337.3,153.3c-7.4,11.3-14.3,15.2-24,15.2c-12.3,0-24.6-8.9-24.6-29.7c0-18,11.1-29.4,25-29.4c21.1,0,22.7,21.1,23,27.7h-30.6
-	c0.3,10.3,3,22.7,14.5,22.7c7.1,0,11.4-5.4,13.7-8.1L337.3,153.3z"
-    />
-    <path
-      style={styles.blue}
-      d="M579,43c-7.2,0-9.7,6.1-18.8,3.3c-0.8-0.3-1.5,0.7-0.8,1.3c5.5,4.6,14.1,11.2,24,10.6
-	C591.3,57.6,590.7,43,579,43"
-    />
-    <path
-      style={styles.pink}
-      d="M588.3,63.8c-0.5,2.5-1.3,5.5-2,7.4c-0.2,0.4-0.7,0.6-1.1,0.4c-18.2-8.8-38.9,11.7-60.7,5.1
-	c-17.7-5.3-18.7-40,3.9-40.5c18.3-0.4,32.8,30.4,59,26.6C587.9,62.7,588.4,63,588.3,63.8"
-    />
-    <path
-      style={styles.pink}
-      d="M533,94.4c9.5,1.2,22.8,8.9,30.8,9.3c0.7,0,0.9,0.9,0.4,1.4c-4.7,3.9-9.5,6.3-13.8,6.3c-6.8,0-15-6.1-22.3-15.2
-	C526.9,94.5,531,94.2,533,94.4"
-    />
-    <path
-      style={styles.pink}
-      d="M525.2,90.4c5.4-3.4,16.5,0.4,27.6,4.9c10.6,4.3,16.1,4.5,20.3,0.3c4.8-4.8,14.5-18.8,7.5-20.6
-	c-16.1-4.1-38.2,15.1-60.6,5.3c-0.7-0.3-1.3,0.3-1,1c0.4,1.1,2.8,6.1,4.9,8.9C524.3,90.6,524.6,90.8,525.2,90.4"
-    />
-    <path
-      style={styles.pink}
-      d="M541.8,15.1c6,0,6.8,8.1,13.1,15.4l0,0c0.5,0.7-0.1,1.3-0.5,1.3C534.4,31,533.4,15.1,541.8,15.1"
-    />
-    <path
-      style={styles.pink}
-      d="M560.4,28.9c-1.1-0.7-2.3-1.8-3.2-3.2c-0.2-0.2-0.2-0.5,0-0.8c1.8-2.9,3.4-6.7,5.8-8.3
-	c3.7-2.8,10.6,0.6,7.8,6.2c-1.1,2.2-3.6,4.5-6.2,6.1C562.8,30,561.2,29.5,560.4,28.9"
-    />
-  </>
-)
-
-export default ({ width }: { width: number }) => (
-  <svg width={width} viewBox="0 0 612 205.1">
-    {inner}
-  </svg>
-)
