--- src/components/ExchangePage/logos/paybis.js
@@ -1,46 +1,0 @@
-// @flow
-
-import React, { Fragment } from 'react'
-
-const inner = (
-  <Fragment>
-    <g fill="none" fillRule="evenodd">
-      <path d="M18.615 13.763h24.359V38.91h-24.36z" />
-      <path
-        fill="#424644"
-        d="M27.15 26.634h2.682c1.626 0 2.927-.297 3.902-.946 1.003-.595 1.599-1.46 1.843-2.596.243-1.298.054-2.19-.596-2.731-.65-.54-1.897-.784-3.74-.784H28.56l-1.41 7.057zm-1.572-12.87h3.577l-.623 3h3.062l.596-3h3.55l-.597 3h1.3c2.493 0 4.255.542 5.339 1.623 1.056 1.055 1.381 2.65.975 4.705-.406 2.028-1.355 3.596-2.9 4.705-1.517 1.109-3.522 1.65-5.933 1.65H26.58l-.271 1.297-1.517 7.599v.027l-.136.513h-5.771c.054-.162.081-.351.135-.54l3.306-16.603c.19-.892.135-1.514-.136-1.812-.298-.297-.948-.46-1.95-.46h-1.626l.515-2.703h5.852l.596-3.002z"
-      />
-      <path d="M.813-.027H127.35v56.621H.813z" />
-      <path
-        fill="#3BB0DF"
-        d="M35.441 33.502c0 .081-.027.163-.027.244-.27 1.298-.84 2.19-1.707 2.73-.84.542-2.195.812-4.037.812h-.57c-3.711 0-9.537.081-13.276.081l-1.328 2.704h5.826l-.542 2.731h3.55l.541-2.73h3.062l-.542 2.73h3.55l.542-2.73h2.628c2.439 0 4.417-.542 5.934-1.596 1.49-1.082 2.438-2.65 2.872-4.732.406-2.028.054-3.57-1.03-4.678l-.054-.054 1.68-1.38h-7.451l-1.49 7.41 1.87-1.542zm8.02 17.008a26.402 26.402 0 0 1-14.333 4.219c-14.632 0-26.473-11.817-26.473-26.445 0-14.629 11.841-26.445 26.473-26.445 10.025 0 19.13 5.624 23.654 14.547h2.06C50.208 6.408 40.182 0 29.128 0 13.494 0 .813 12.655.813 28.284s12.68 28.283 28.315 28.283a28.4 28.4 0 0 0 17.53-6.057h-3.197zm82.182-16.845c.65-2.894-1.572-3.678-3.116-4.273-.949-.351-1.653-.676-1.463-1.541.135-.649.84-1.109 1.842-1.109.867 0 2.005.433 2.52.758l1.897-2.84c-.976-.595-2.385-1-4.282-1-3.414 0-5.744 1.622-6.34 4.218-.623 2.84 1.572 3.678 3.089 4.272.921.352 1.626.703 1.436 1.488-.162.703-.867 1.162-1.842 1.162-.867 0-2.222-.378-2.873-.757l-.84 3.624c.895.27 1.951.432 3.09.432 3.928 0 6.286-1.703 6.882-4.434zm-12.6 4.137l3.116-13.845h-4.227l-3.116 13.845h4.227zm4.281-17.928c.271-1.217-.596-2.217-1.87-2.217-1.3 0-2.573 1-2.844 2.217-.271 1.217.569 2.19 1.842 2.19 1.3 0 2.601-.973 2.872-2.19zm-12.49 10.924c-.57 2.57-1.925 4.03-3.686 4.03-.488 0-.948-.082-1.192-.19l1.653-7.382c.38-.162.921-.297 1.409-.297 1.707 0 2.357 1.406 1.815 3.84zm4.335-.378c.948-4.218-.976-6.652-4.58-6.652-.731 0-1.68.19-2.276.433l1.463-6.571H99.55l-4.39 19.469c1.03.54 2.9 1 4.959 1 4.714 0 7.884-2.488 9.05-7.68z"
-      />
-      <path d="M69.853 24.2H97.03v17.334H69.853z" />
-      <path
-        fill="#424644"
-        d="M96.975 24.471H92.45l-3.522 7.571a24.423 24.423 0 0 0-.894 2.217h-.027s.081-1.19 0-2.217l-.57-7.57h-4.524l2.384 13.167-2.222 3.867h4.227l9.673-17.035zM80.61 36.963l1.544-7.922c.677-3.488-.867-4.84-4.444-4.84-1.03 0-2.113.108-3.034.297-.271 1.298-.813 2.38-1.653 3.218 1.002-.352 2.466-.703 3.658-.703 1.355 0 1.978.487 1.788 1.541l-.108.514c-4.038.216-7.75 1.027-8.427 4.515-.515 2.786 1.22 4.354 5.338 4.354 2.14 0 4.254-.433 5.338-.974zm-3.387-2.028c-.271.109-.976.19-1.518.19-1.436 0-2.222-.568-2.032-1.65.325-1.54 2.005-1.784 4.2-1.892l-.65 3.352z"
-      />
-      <path d="M53.704 19.144H73.51v18.522H53.704z" />
-      <path
-        fill="#424644"
-        d="M60.749 27.364h1.87c1.354 0 2.438-.27 3.25-.757.813-.514 1.328-1.244 1.518-2.19.217-1.082.054-1.839-.488-2.271-.542-.46-1.571-.676-3.089-.676h-1.896l-1.165 5.894zm7.37-8.22c2.059 0 3.522.433 4.389 1.325.894.92 1.165 2.217.84 3.948-.352 1.676-1.138 3.001-2.412 3.894-1.273.946-2.899 1.406-4.93 1.406h-5.718l-.217 1.081-1.246 6.355-.027.027-.082.432H53.92c.055-.135.082-.297.109-.46l2.764-13.844c.135-.784.108-1.27-.136-1.54-.217-.244-.759-.38-1.626-.38h-1.327l.433-2.244H68.12z"
-      />
-      <path d="M50.344 42.94h75.895v6.813H50.344z" />
-      <path
-        fill="#424644"
-        d="M53.27 48.023l1.3-5.084h-1.11l-.434 1.677a1.544 1.544 0 0 0-.433-.081c-1.111 0-1.897.73-2.168 1.892-.298 1.19.163 1.866 1.382 1.866.569 0 1.138-.108 1.463-.27zm-.948-.622c-.054.027-.217.027-.325.027-.461 0-.624-.352-.461-1.028.163-.622.488-1.027.975-1.027.109 0 .271.027.326.054l-.515 1.974zm4.01-3.894c.081-.324-.136-.568-.488-.568-.325 0-.677.244-.758.568-.082.325.135.568.487.568.325 0 .678-.243.759-.568zm-1.274 4.705l.922-3.623h-1.084l-.948 3.623h1.11zm3.902-.081l.867-3.38c-.27-.135-.731-.243-1.273-.243-1.22 0-2.087.622-2.412 1.838-.27 1.082.19 1.73 1.111 1.73.217 0 .488-.053.65-.107l-.027.162c-.108.433-.379.649-.84.649-.433 0-.785-.081-1.002-.19l-.515.839c.298.162.759.297 1.382.297 1.002 0 1.788-.54 2.06-1.595zm-.84-.946c-.108.027-.27.08-.406.08-.407 0-.596-.324-.434-.973.163-.649.515-.946.949-.946.135 0 .27 0 .352.027l-.46 1.812zm3.93-3.678c.08-.324-.136-.568-.489-.568-.325 0-.677.244-.758.568-.082.325.135.568.487.568.326 0 .678-.243.76-.568zm-1.274 4.705l.92-3.623h-1.083l-.948 3.623h1.11zm2.71-.054l.243-.92a.976.976 0 0 1-.298.055c-.217 0-.325-.081-.271-.325l.406-1.568h.624l.216-.811h-.65l.271-1.082-1.138.298-.217.784h-.406l-.217.81h.434l-.434 1.65c-.19.812 0 1.217.759 1.217.27 0 .542-.054.677-.108zm3.603-.135l.569-2.19c.244-.974-.163-1.325-1.138-1.325a4.81 4.81 0 0 0-1.517.243l.08.73c.299-.108.76-.216 1.085-.216.379 0 .569.135.487.432l-.054.136c-1.11.054-2.14.297-2.384 1.27-.19.785.27 1.19 1.409 1.19.569 0 1.165-.108 1.463-.27zm-.894-.541a1.471 1.471 0 0 1-.407.054c-.406 0-.596-.162-.542-.46.109-.432.597-.486 1.193-.513l-.244.919zm2.71.73l1.327-5.273h-1.11l-1.329 5.273h1.111zm11.217-.162l-.136-.811a1.04 1.04 0 0 1-.514.135c-.38 0-.596-.379-.434-.974.163-.649.515-1 .948-1 .19 0 .353.027.434.135l.542-.784c-.19-.162-.46-.243-.921-.243-.895 0-1.816.594-2.14 1.892-.326 1.271.27 1.893 1.164 1.893.46 0 .786-.081 1.057-.243zm3.522-3.515c-.135-.027-.298-.027-.514-.027-.65 0-1.193.08-1.653.243l-.867 3.461h1.11l.705-2.785c.108-.054.244-.054.38-.054.162 0 .297.027.406.054l.433-.892zm4.038.054h-1.247l-1.084 2.082c-.162.297-.298.622-.298.622s.027-.325.027-.622l-.027-2.082h-1.246l.433 3.623-.894 1.433h1.138l3.198-5.056zm3.251 1.811c.298-1.243-.162-1.892-1.382-1.892-.569 0-1.138.108-1.49.243l-1.246 4.894h1.11l.407-1.46c.081.027.27.054.433.054 1.111 0 1.87-.73 2.168-1.839zm-1.11 0c-.163.595-.516 1.028-.976 1.028a.886.886 0 0 1-.325-.054l.514-1.974a.886.886 0 0 1 .326-.054c.46 0 .623.351.46 1.054zm3.17 1.758l.243-.92a.976.976 0 0 1-.298.055c-.217 0-.325-.081-.27-.325l.406-1.568h.623l.217-.811h-.65l.297-1.082-1.165.298-.216.784h-.407l-.217.81h.434l-.434 1.65c-.19.812 0 1.217.759 1.217.27 0 .542-.054.677-.108zm4.253-1.758c.271-1.108-.135-1.892-1.219-1.892s-1.897.784-2.168 1.892c-.298 1.082.136 1.893 1.193 1.893 1.083 0 1.923-.811 2.194-1.893zm-1.11 0c-.163.65-.461 1-.84 1-.407 0-.515-.35-.326-1 .163-.703.461-1.054.84-1.054.38 0 .515.351.325 1.054zm5.419 1.65l-.136-.811a1.04 1.04 0 0 1-.515.135c-.379 0-.596-.379-.433-.974.162-.649.515-1 .921-1 .19 0 .352.027.46.135l.543-.784c-.217-.162-.461-.243-.949-.243-.867 0-1.815.594-2.14 1.892-.325 1.271.298 1.893 1.192 1.893.46 0 .759-.081 1.057-.243zm3.63-.027l.895-3.434h-1.111l-.705 2.758a1.701 1.701 0 0 1-.433.054c-.325 0-.434-.135-.38-.46l.624-2.352h-1.111l-.597 2.325c-.243.974.19 1.38 1.247 1.38.569 0 1.192-.109 1.571-.271zm3.713-3.488c-.109-.027-.299-.027-.515-.027-.623 0-1.192.08-1.626.243l-.894 3.461h1.11l.705-2.785c.109-.054.244-.054.407-.054.135 0 .298.027.379.054l.434-.892zm2.6 0c-.135-.027-.324-.027-.514-.027-.65 0-1.22.08-1.653.243l-.867 3.461h1.11l.705-2.785c.109-.054.244-.054.38-.054.162 0 .298.027.406.054l.434-.892zm3.144 1.65c.244-.974-.136-1.677-1.192-1.677-1.084 0-1.951.757-2.222 1.892-.271 1.082.135 1.893 1.382 1.893.542 0 1.138-.108 1.49-.27l-.108-.811a2.732 2.732 0 0 1-1.057.216c-.46 0-.759-.19-.732-.568l2.33-.297c.028-.136.055-.27.109-.379zm-1.03-.217l-1.3.162c.135-.595.514-.865.894-.865s.515.243.406.703zm4.254 2.244l.623-2.488c.217-.784-.162-1.216-1.3-1.216-.65 0-1.138.08-1.599.243l-.867 3.461h1.111l.705-2.785a1.89 1.89 0 0 1 .433-.054c.325 0 .46.162.38.46l-.597 2.379h1.111zm3.306-.162l-.136-.811a1.04 1.04 0 0 1-.514.135c-.38 0-.597-.379-.434-.974.163-.649.515-1 .921-1 .19 0 .353.027.46.135l.543-.784c-.217-.162-.46-.243-.948-.243-.868 0-1.816.594-2.141 1.892-.298 1.271.298 1.893 1.192 1.893.46 0 .759-.081 1.057-.243zm5.013-3.461h-1.247l-1.11 2.082c-.136.297-.272.622-.272.622s.027-.325.027-.622l-.027-2.082h-1.246l.433 3.623-.92 1.433h1.164l3.198-5.056zM76.03 45.67l.082-.838-1.897.325c-.19.027-.325.027-.433.027-.38 0-.57-.19-.488-.514.081-.298.352-.487.786-.487.27 0 .542.081.704.19l.542-.866c-.244-.135-.65-.243-1.165-.243-1.057 0-1.761.622-1.951 1.325-.108.432.027.838.434 1.027-.515.19-1.03.595-1.193 1.271-.216.92.38 1.406 1.626 1.406.677 0 1.274-.108 1.626-.27l.569-2.245.758-.108zm-2.22 1.73a2.435 2.435 0 0 1-.488.055c-.57 0-.813-.216-.705-.649.109-.406.434-.676.922-.757l.623-.081-.352 1.433z"
-      />
-      <path d="M50.317 49.78h29.399v6.814h-29.4z" />
-      <path
-        fill="#424644"
-        d="M53.812 53.025c.244-.973-.135-1.676-1.192-1.676-1.084 0-1.924.757-2.222 1.892-.244 1.082.135 1.893 1.409 1.893.542 0 1.11-.108 1.49-.27l-.108-.811a2.766 2.766 0 0 1-1.084.216c-.434 0-.732-.19-.732-.568l2.33-.297c.055-.135.082-.27.109-.379zm-1.003-.216l-1.327.162c.162-.595.542-.865.921-.865.352 0 .515.243.406.703zm4.471 2.244l-.758-1.866 1.571-1.757h-1.246l-.813 1.027h-.027l-.271-1.027h-1.274l.705 1.757-1.707 1.866h1.246l.949-1.136.352 1.136h1.273zm2.791-.162l-.135-.811a1.04 1.04 0 0 1-.515.135c-.407 0-.596-.379-.434-.974.163-.648.515-1 .922-1 .19 0 .352.027.46.135l.542-.784c-.217-.162-.488-.243-.948-.243-.867 0-1.816.595-2.14 1.892-.326 1.271.297 1.893 1.191 1.893.461 0 .76-.08 1.057-.243zm3.63.162l.624-2.46c.19-.812-.162-1.244-1.03-1.244-.243 0-.514.08-.731.162l.433-1.73h-1.11l-1.355 5.272h1.11l.705-2.758a1.23 1.23 0 0 1 .434-.081c.352 0 .487.162.406.487l-.596 2.352h1.11zm3.74-.19l.57-2.19c.243-.973-.163-1.324-1.166-1.324-.542 0-1.138.108-1.49.243l.081.73c.298-.108.732-.216 1.084-.216.38 0 .542.135.46.432l-.027.136c-1.11.054-2.167.297-2.411 1.27-.19.758.244 1.19 1.436 1.19.569 0 1.165-.108 1.463-.27zm-.894-.54a1.643 1.643 0 0 1-.406.054c-.407 0-.624-.162-.542-.46.108-.432.569-.486 1.192-.513l-.244.92zm4.715.73l.65-2.488c.19-.784-.19-1.216-1.328-1.216-.65 0-1.138.08-1.571.243l-.894 3.461h1.11l.705-2.785a1.89 1.89 0 0 1 .433-.054c.326 0 .461.162.38.46l-.596 2.38h1.11zm3.901-.081l.84-3.38c-.27-.135-.704-.243-1.273-.243-1.192 0-2.086.622-2.385 1.838-.27 1.082.19 1.73 1.111 1.73.217 0 .461-.053.65-.107l-.054.162c-.108.433-.379.649-.84.649-.406 0-.758-.081-.975-.19l-.542.839c.325.162.759.297 1.41.297 1.002 0 1.787-.54 2.058-1.595zm-.867-.946a1.324 1.324 0 0 1-.379.08c-.434 0-.623-.324-.46-.973.162-.649.541-.946.948-.946.162 0 .27 0 .352.027l-.46 1.812zm5.311-1c.244-.974-.135-1.677-1.192-1.677-1.084 0-1.924.757-2.222 1.892-.27 1.082.136 1.893 1.409 1.893.542 0 1.111-.108 1.49-.27l-.108-.811a2.87 2.87 0 0 1-1.084.216c-.46 0-.731-.19-.731-.568l2.33-.297c.054-.135.081-.27.108-.379zm-1.002-.217l-1.328.162c.162-.595.515-.865.921-.865.352 0 .515.243.407.703z"
-      />
-    </g>
-  </Fragment>
-)
-
-export default ({ width, height }: { width: number, height: number }) => (
-  <svg width={width} height={height}>
-    {inner}
-  </svg>
-)
